+++ src/job/jigdodownload.hh
@@ -1,4 +1,4 @@
-/* $Id: jigdodownload.hh,v 1.7 2003-08-15 11:38:30 atterer Exp $ -*- C++ -*-
+/* $Id: jigdodownload.hh,v 1.8 2004-06-24 10:13:34 atterer Exp $ -*- C++ -*-
   __   _
   |_) /|  Copyright (C) 2003  |  richard@
   | \/¯|  Richard Atterer     |  atterer.net
@@ -56,7 +56,7 @@ private:
   virtual void job_failed(string* message);
   virtual void job_message(string* message);
   virtual void dataSource_dataSize(uint64 n);
-  virtual void dataSource_data(const byte* data, size_t size,
+  virtual void dataSource_data(const byte* data, unsigned size,
                                uint64 currentSize);
 
   // Virtual methods from Gunzip::IO
